working on turning collection components into functional components, still need to refactor a bit
